Notepad Automation for Win 10
Notepad dialogs and windows style has change in Windows 10. 
Old example works in Windows 7 but not in Windows 10. This file will specifically work on Windows 10 only.
In this program Notepad will be opened and it will first open Page Setup Dialog and after finishing the jobs it will open Print Dialog.

In the end it will write some text in Notepad and then the program will exit.