Enable finding of undisplayed menu items
I found with our UIA app that:

config_item = menu.item_by_path("File -> Configuration", False)

did not work. Pywinauto crashes in fact. Traced to the fact that menu.items() does not see the File menu's items until the File menu is clicked and they are displayed. That is the UIA controls don't exist until the they are displayed. 

This line of code works with this patch, and I can find and then click the configuration item on the FIle menu using:

config_item.click_input()

and all is good.