enable `cache_enable` in 'create_tree_walker' mode
In pywinauto/pywinauto#207, the `cache_enable` argument was added to the
constructor, propagating to `children` and `descendants`. However, when
`iter_children` and `iter_descendants` were introduced in
pywinauto/pywinauto#475, `cache_enable` wasn't implemented to propagate.

From reviewing past issues and PRs, it appears that not propagating
`cache_enable` was unintentional and simply overlooked.
pywinauto/pywinauto#1001 and pywinauto/pywinauto#1012 refactored the
code without changing the current behavior, resulting in a complex
conditional implementation.